Move `mediatypes` to `app.toolkit`
