+++ app/app/blobs/services/thumbnailer/thumbnailer.py
@@ -4,9 +4,10 @@ import os.path
 from typing import IO, TYPE_CHECKING
 
 from app.app.blobs.domain.content import InMemoryBlobContent
-from app.app.files.domain import File, mediatypes
+from app.app.files.domain import File
 from app.app.files.services.thumbnailer import thumbnails
 from app.cache import cache
+from app.toolkit import mediatypes
 
 if TYPE_CHECKING:
     from collections.abc import Iterable
