+++ app/app/files/services/metadata/readers/__init__.py
@@ -3,7 +3,8 @@ from __future__ import annotations
 import asyncio
 from typing import IO
 
-from app.app.files.domain import Exif, mediatypes
+from app.app.files.domain import Exif
+from app.toolkit import mediatypes
 from app.toolkit.mediatypes import MediaType
 
 from .image import load_image_data
