+++ app/app/files/services/thumbnailer/thumbnails/__init__.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import asyncio
 from typing import IO
 
-from app.app.files.domain import mediatypes
+from app.toolkit import mediatypes
 from app.toolkit.mediatypes import MediaType
 
 from .image import thumbnail_image
